기본적인 회원가입, 로그인 기능 개발

See merge request s13-fintech-finance-sub1/S13P21A408!11

+++ Backend/src/main/java/com/mango/backend/domain/user/service/UserService.java
@@ -2,12 +2,12 @@ package com.mango.backend.domain.user.service;
 
 import com.mango.backend.domain.user.entity.User;
 import com.mango.backend.domain.user.repository.UserRepository;
-import com.mango.backend.global.common.api.BaseResponse;
-import com.mango.backend.global.common.api.ErrorResponse;
-import com.mango.backend.global.common.api.SuccessResponse;
+import com.mango.backend.global.common.ServiceResult;
 import com.mango.backend.global.error.ErrorCode;
+import com.mango.backend.global.util.JwtProvider;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,16 +17,26 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
   private final UserRepository userRepository;
+  private final RedisTemplate<String, String> redisTemplate;
+  private final JwtProvider jwtProvider;
 
   @Transactional
-  public BaseResponse deleteUser(Long userId) {
+  public ServiceResult<Void> deleteUser(Long userId, String token) {
+
+    Long requestId = jwtProvider.getUserId(token);
+    if (!userId.equals(requestId)) {
+      return ServiceResult.failure(ErrorCode.AUTH_FORBIDDEN);
+    }
+
     Optional<User> userOpt = userRepository.findById(userId);
     if (userOpt.isEmpty()) {
-      return ErrorResponse.of(ErrorCode.USER_INVALID_INPUT);
+      return ServiceResult.failure(ErrorCode.USER_INVALID_INPUT);
     }
 
     userRepository.delete(userOpt.get());
-    return SuccessResponse.of("회원 탈퇴 성공", null);
-  }
+    String redisKey = "JWT:" + userId;
+    redisTemplate.delete(redisKey);
 
+    return ServiceResult.success(null);
+  }
 }
